json parsing

+++ src/audio.h
@@ -9,7 +9,6 @@
 //   gcc  : -lasound
 
 #include <stdbool.h>
-#include <stdint.h>
 #ifdef _WIN32
 #  include <xaudio2.h>
 #elif linux
@@ -40,7 +39,7 @@ typedef struct{
 
 // Public
 AUDIO_DEF bool audio_init(Audio *audio, Audio_Fmt fmt, int channels, int sample_rate);
-AUDIO_DEF void audio_play(Audio *audio, unsigned char *data, uint64_t samples);
+AUDIO_DEF void audio_play(Audio *audio, unsigned char *data, int samples);
 //AUDIO_DEF void audio_play_async(Audio *audio, unsigned char *data, int samples);
 AUDIO_DEF void audio_block(Audio *audio);
 AUDIO_DEF void audio_free(Audio *audio);
@@ -164,10 +163,10 @@ AUDIO_DEF bool audio_init(Audio *audio, Audio_Fmt fmt, int channels, int sample_
     return true;
 }
 
-AUDIO_DEF void audio_play(Audio *audio, unsigned char *data, uint64_t samples) {
+AUDIO_DEF void audio_play(Audio *audio, unsigned char *data, int samples) {
     XAUDIO2_BUFFER xaudioBuffer = {0};
   
-    xaudioBuffer.AudioBytes = (UINT32) (samples * audio->sample_size);
+    xaudioBuffer.AudioBytes = samples * audio->sample_size;
     xaudioBuffer.pAudioData = data;
     xaudioBuffer.pContext = audio;
     
@@ -176,10 +175,10 @@ AUDIO_DEF void audio_play(Audio *audio, unsigned char *data, uint64_t samples) {
     WaitForSingleObject((audio)->semaphore, INFINITE);
 }
 
-AUDIO_DEF void audio_play_async(Audio *audio, unsigned char *data, uint64_t samples) {
+AUDIO_DEF void audio_play_async(Audio *audio, unsigned char *data, int samples) {
     XAUDIO2_BUFFER xaudioBuffer = {0};
   
-    xaudioBuffer.AudioBytes = (UINT32) samples * audio->sample_size;
+    xaudioBuffer.AudioBytes = samples * audio->sample_size;
     xaudioBuffer.pAudioData = data;
     xaudioBuffer.pContext = audio;
     
@@ -263,7 +262,7 @@ AUDIO_DEF bool audio_init(Audio *audio, Audio_Fmt fmt, int channels, int sample_
   return true;  
 }
 
-AUDIO_DEF void audio_play(Audio *audio, unsigned char *data, uint64_t samples) {
+AUDIO_DEF void audio_play(Audio *audio, unsigned char *data, int samples) {
 
   while(samples > 0) {
     int ret = snd_pcm_writei(audio->alsa_snd_pcm, data, samples);
